F2U file/dir selection and conversion works! Some restruct (F2UConvPanel code -> ConvPanel) needs to be done still

+++ ConversionTools4/src/np/org/mpp/conv4/utils/jodconverter/AnyDocumentChainedReaderWriter.java
@@ -63,8 +63,10 @@ public class AnyDocumentChainedReaderWriter implements GeneralReaderWriter {
 
       if (!oooSupportedExtentions.contains(outExt)) {
         if (supportedExtentions.contains(outExt)) {
-          jodDocumentConverter.convert(new File(convOutFile), new File(outFile));
-          new File(convOutFile).delete();
+          if (new File(convOutFile).exists()) { // file may not be there is there was no changes
+            jodDocumentConverter.convert(new File(convOutFile), new File(outFile));
+            new File(convOutFile).delete();
+          }
         } else {
           // out file format not supported. Just rename and throw error
           String outFile2 = FilenameUtils.removeExtension(outFile) + "." + FilenameUtils.getExtension(convOutFile);
